Conflicts:
	src/server/game/Achievements/AchievementMgr.cpp
	src/server/game/DataStores/DBCEnums.h
	src/server/game/Handlers/PetHandler.cpp
	src/server/game/Server/WorldSession.h

+++ src/server/game/Handlers/PetHandler.cpp
@@ -143,7 +143,7 @@ void WorldSession::HandlePetStopAttack(WorldPacket &recvData)
     pet->AttackStop();
 }
 
-void WorldSession::HandlePetActionHelper(Unit* pet, uint64 guid1, uint16 spellid, uint16 flag, uint64 guid2, float x, float y, float z)
+void WorldSession::HandlePetActionHelper(Unit* pet, uint64 guid1, uint32 spellid, uint16 flag, uint64 guid2, float x, float y, float z)
 {
     CharmInfo* charmInfo = pet->GetCharmInfo();
     if (!charmInfo)
